backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
python-dotenv==1.0.0
httpx==0.25.2
tweepy==4.14.0
langchain==0.0.350
langchain-openai==0.0.2
openai==1.3.7
beautifulsoup4==4.12.2
aiofiles==23.2.1
python-multipart==0.0.6
jinja2==3.1.2
pytest==7.4.3
pytest-asyncio==0.21.1